Reshuffle androidx dependencies not to require them being listed… (#397)
Unfortunately we still need to require swipetorefresh as some of the material classes use it without listing swipetorefresh package as dependency. However it appears like we can put all the dependencies in the screen build.gradle and not require additional dependencies to be listed in the app's top level gradle.