Fix push updates during transition (#409)
This change fixes an issue that was caused by us updating view controllers while there was an ongoing transition.

When changes to viewControllers where made during other transition, the navigation controller wouldn't update view controllers property immediately. As a result we could've get out of sync between stack subviews and ciewController property. And some of the stack logic relied on those two props being in sync. In particular, when we make a push transition we expected viewControllers to contain newly pushed item so that we won't push it again in the future. However, the item wouldn't be added immediately if the controller was running the transition.

To fix that there are two key changes introduced here:
1) when updating push controllers, we check if nav controller is already running a transition. If so, we wait with the updates until the transition is over.
2) we no longer use didShowViewController delegate method to mark screens as dismissed. Apparently, that approach does not work when viewContollers and subviews are out of sync. Because of it, we'd mark newly added screens as dismissed just because there weren't yet added to viewControllers prop. Instead we introduced dismissed property directly in screen class which is set in viewWillDisappear. This way we no longer need to keep a set of dismissed screens.