Fix screen container to prevent event mutex lock (#473)
This change fixes RNSScreenContainer on iOS to prevent from reentering RN's event emitter which resulted in app not responding when used together with react-navigation v5.

The original issue could've been reproduced when using simple react-navigation stack with two views and when attempting to swipe back from the detail view back to home. When that happened the app would go into a deadlock and stop responding.

The underlying issue was due to the fact that we'd trigger detach for screens that are going to be kept active just to make sure that screens are kept in the correct order. As a result, detach would trigger gesture handler to cancel swipe which as a result would attempt to generate an event. Since this whole process would already happen as a result of gesture handler event (during ongoing pan gesture), the RN's gesture emitter would go into deadlock as it uses mutex w/o reentry flag.

As we don't need to go through the full detach/attach lifecycle for screens that are kept this change updates that logic to only dispatch willMoveToParent for screens that are active at the time when new active screen is added. We also had to update attachScreen method in order to support index at which the new screen should be placed.

On top of the above I also noticed that there was a second problem caused by a recent change landed in #378, which would result in the screen we removed from the react hierarchy to not be detached. This was due to the fact we were relying on accessing controller reference via RNSScreenView (view.controller) while the controller ref was being reset to avoid reference cycles. In order to fix this problem we added findChildControllerForScreen which looks for the screen's controller reference in childViewControllers array instead of accessing controller property in RNSScreenView.