Fix user interface style in native stack view controllers (#481)
# Why

In Expo client, which supports both light and dark styles, you can open an experience that has strict setting to light or dark mode. In that case experience's root view controller is set to the style specified by the user, however view controllers presented over it or added as a child have `overrideUserInterfaceStyle` set to `UIUserInterfaceStyleUnspecified` which causes the view controller to use the style that is set in system settings so we get some UIKit's views rendered with incorrect interface style 😞 This only affects view controllers presented modally, thus doing the same for pushed screens is not necessary.
@brentvatne prepared reproducible example: https://github.com/brentvatne/modal-example

# How

Native stack's view controllers will inherit `overrideUserInterfaceStyle` from presenting view controller.

# Test Plan

Example https://github.com/brentvatne/modal-example works as expected in Expo client with style set to Dark in system settings. _*This also requires one small change in the client itself, so it won't work in the current version in the app store._