Add props to headerLeft and headerRight (#500)
This PR will allow users to custom headerRight and headerLeft components using user's theme color.
For example, if I want a custom settings icon in headerRight to navigate to my settings page, I will be able to retrieve color from my custom theme. It's already available in ReactNavigation.

colors.primary is used by the back button so it makes sense to use it for headerRight, headerLeft and headerCenter.