fix: canceled swipe gesture on iOS 12 when direction hasn't changed (#1150)
Setting up the default value for direction in native stack v6 revealed an issue with a swipe gesture on iOS 12. Setting direction with the same value twice cancels the transition between screens in iOS 12. This issue doesn't occur on iOS 13+ because Apple reimplemented how the gesture system works. This PR ensures the direction is set only when needed.