fix: android keyboard with autofocus (#1214)
Change the logic of preventing Android keyboard dismiss in views with autofocus prop. We now prevent the keyboard from hiding by removing the call to clearFocus at the beginning of fragment transition, which is the reason why the keyboard is hidden almost exactly after it is shown on the new screen.