feat: bring Fabric to SearchBar (#1470)
* chore: add JS for SearchBar

Component style props still have to be adjusted. Same for other props
that are not implemented yet.

* chore: add event types to codegen

* chore: add iOS props to JS

* chore: add Android only props to JS

* chore: update interface for Fabric

* chore: add enum conversion for autoCapitalization prop

* chore: add props update (except colors)

* chore: add componentDescriptorProvider & RNSSearchBarCls

* chore: make event blocks Paper specific (in interface)

* chore: make Fabric impl not emit any events (just for now)

* fix: import ComponentDescriptors header

* refact: extract common initialization code to initCommonProps

* chore: create SearchButtonPressedEvent type

* chore: create ChangeTextEvent type

* chore: create event emitting methods

* refact: make use of event emitting methods

* chore: move #import RNSSearchBar.h statemets to common section

This can be done now, as SearchBar should work on Fabric

* refact: add TODO for missing props

* chore: change ctor for Fabric in RNSSearchBarManager

I don't exactly know if it is required but it should not break anything.
If this change is useless we can restore it later.

* chore: make RNSSearchbarManager#view method paper only

* refact: update TODO comments

* chore: move RNSScreen#hideHeaderIfNecessary method to common section

This change was not tested. Usage of self.view.reactSubviews raises my
concerns as I'm not sure if this property always holds correct state on
Fabric

* chore: remove message that SearchBar is not yer Fabric compatible

* chore: udpate style for SearchBar component

* feat: implement color-related props

* chore: enable call to hideheaderIfNecessary

* chore: add Test758 from TestExample

* fix: weird merge artifact

This if def was present on main & not on current branch, but after running `git merge main` git took the version of code from the current branch?!

* chore: add Test758 to App.js