fix(iOS): select correct VC for nested modal presentation (#1912)
## Description

Closes #1829

Currently it is not possible to navigate to a modal from another modal
view that is mounted in JS under nested stack (see #1829 description for
better context).
This is the case, because modal is presented by
`RNSNavigationController` corresponding to the stack under which it is
mounted in JS ==> UIKit reports that such controller is already
presenting.

IMO the last presented modal should be the one, to present new one.


## Changes

Added code handling above case, by checking whether modals from other
stacks are present & using top-most modal for presentation in such case.

## Test code and steps to reproduce

See `Test1829` in test example apps

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes (merged only comments after final CI
checks)