fix(iOS): app freeze when navigating back from any modal nested in contained modal (#1996)
## Description

Updated logic in `RNSScreenStackView#setModalViewControllers:` so that
it handles case when navigating back from top-level VC that was not
presented
by the screen stack itself, as it happens with
`UIModalPresentationCurrentContext` &
`UIModalPresentationOverCurrentContext`.

Fixes #1813

## Changes

We now check whether the top-level VC should be dismissed by the stack
being asked.

## Test code and steps to reproduce

See #1813 for reproduction steps.

I'm also adding these in `Example` application.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes