fix(iOS): incorrect safe area on transparent modals using landscape orientation (#2008)
## Description
Normally, while opening a modal, each screen should preserve its safe
area, depending on the screen orientation. Unfortunately, when user
dismisses transparent modals, safe area of the origin screen breaks,
resulting in the vertical safe area on landscape orientation.
Because of that, I investigated the logic behind the final decision of
screen orientations and it looks like the problem was lying on
`supportedScreenOrientations` method in RNSScreen.mm file.

1. First, the modal was being asked for its supported device
orientation. Once it reached the `supportedScreenOrientations` method,
it was asking for config of childVC. Because the childVC was `nil` and
the orientation wasn't set in screen options, it was returning `nil`,
resulting in returning all orientations without `upside down`.
2. After that, there was a time for looking onto child VCs of a screen
behind the modal. Since it was presenting a modal, it was first checking
for its last child - since it didn't provide any modal and it haven't
got any children, the last child was also `nil`, resulting in returning
the modal as an orientation of the screen.
3. Returning a modal is (probably) a bad idea here, since it does not
have any screen orientation specified, resulting in returning `nil` as a
screen orientation. This was probably resulting a bug with wrong safe
area.

This PR changes this bad behavior by not returning `lastViewController`
(which is a modal that is being presented from a screen) and going
further for looking a config in child view controllers.
However, this behavior may lead to the further bugs we haven't
discovered yet.

## Changes

- Changed `return` statement in `findChildVCForConfigAndTrait` method.

## Screenshots / GIFs

### Before


https://github.com/software-mansion/react-native-screens/assets/23281839/6ff888d9-a7df-466c-ac45-d3db1cbe928c

### After


https://github.com/software-mansion/react-native-screens/assets/23281839/dfc03a6c-ca7a-49a0-9c3e-7f019f7d405c

## Test code and steps to reproduce

You can check Test2008 in `TestsExample` and `FabricTestExample` in
order to tests how transparent modals behave.

## Checklist

- [X] Included code example that can be used to test this change
- [X] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacper.kafara@swmansion.com>